requirements.txt
flask
opencv-python-headless
piexif
pymediainfo
pillow
matplotlib
geopy
folium